Autogenerate primary key index

+++ iot-device/src/main/resources/mysql/init.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS temperature
 (
     measure_id          INT AUTO_INCREMENT,
-    sensor_id           TINYINT      NOT NULL,
-    current_temperature TINYINT      NOT NULL,
+    sensor_id           INT          NOT NULL,
+    current_temperature INT          NOT NULL,
     status              VARCHAR(255) NOT NULL,
     PRIMARY KEY (measure_id)
 ) ENGINE = INNODB;
